+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add `ExternalAttachment` for externalised attachments ([#372](https://github.com/cucumber/messages/pull/372))
+
+### Changed
+- Deprecate `Attachment.url` and `Attachment.source` ([#372](https://github.com/cucumber/messages/pull/372))
 
 ## [31.2.0] - 2026-01-11
 ### Added