+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Deprecate `Attachment.url` and `Attachment.source` ([#372](https://github.com/cucumber/messages/pull/372))
+- Make StepMatchArgument.Group.children optional ([#370](https://github.com/cucumber/messages/pull/370))
 
 ## [31.2.0] - 2026-01-11
 ### Added