+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [32.0.0] - 2026-01-21
 ### Added
 - Add `ExternalAttachment` for externalised attachments ([#372](https://github.com/cucumber/messages/pull/372))
 
@@ -649,7 +651,8 @@ aggregated status.
 ### Added
 - Protobuf messages for Go, Java, JavaScript, TypeScript and Ruby
 
-[Unreleased]: https://github.com/cucumber/messages/compare/v31.2.0...HEAD
+[Unreleased]: https://github.com/cucumber/messages/compare/v32.0.0...HEAD
+[32.0.0]: https://github.com/cucumber/messages/compare/v31.2.0...v32.0.0
 [31.2.0]: https://github.com/cucumber/messages/compare/v31.1.0...v31.2.0
 [31.1.0]: https://github.com/cucumber/messages/compare/v31.0.1...v31.1.0
 [31.0.1]: https://github.com/cucumber/messages/compare/v31.0.0...v31.0.1